server/requirements.txt
fastapi==0.124.0
PyJWT==2.10.1
aiosqlite==0.21.0
bcrypt==5.0.0
python-multipart==0.0.20
uvicorn==0.38.0

